Add files via upload
Dis a hurted my brain. I thought I had to go through and find out how to run this in all the branches of code (ya know, the subclasses), but I had a version of Java card game where all of the classes weren't separate, so that made it less mentally taxing. However, I only put print statements in the visual part of my code. (The parts where it was mainly the game and not the separate items used to be played.)